Add framework for flexible IO contexts for UM.
* IO interface
* Default IOConsole implementation
* Additional flexible delegate-powered implementation that allows
  the user to populate an IO implementation anonymously.

Details
=======
A UniversalMachine needs the ability to interact with its environment
for exactly three op-codes: Halt, Input, and Output.

Ideally, we'll provide a line-buffered editing environment for use
in normal interactive situations, but in other situations (such as
to enable automatically decompressing the contest UM image), we
don't want to require actual user interaction. (This will also be
critical for our automated solution to the Adventure puzzle.)

To enable any manner of IO contexts for the UM, we define an
interface that requires Halt, Input, and Output implementations. A
default implementation that connects the user directly to the
actual Console is included. Additionally, an class implementing the
IO interface is included whose constructor accepts 3 delegates so
that anonymous implementations may be created for simple needs.